Updating document context
