+++ angular/projects/dynamic-render/src/app/services/document-context.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { DocumentContextService } from './document-context.service';
+
+describe('DocumentContextService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it('should be created', () => {
+    const service: DocumentContextService = TestBed.get(DocumentContextService);
+    expect(service).toBeTruthy();
+  });
+});
