+++ angular/projects/dynamic-render/src/app/services/dynamic-page.service.ts
@@ -60,43 +60,21 @@ export class DynamicPageService {
     this.titleService.setTitle(`${title} | ${siteName}`);
   }
 
-  async render(viewContainer: ViewContainerRef, contentEntry: ContentEntry): Promise<ComponentRef<any>> {
-    const contentType = await this.contentTypeService.getContentType(contentEntry.contentType).pipe(take(1)).toPromise();
+  async render(viewContainer: ViewContainerRef, documentContext: TanamDocumentContext): Promise<ComponentRef<any>> {
+    const contentType = await this.contentTypeService.getContentType(documentContext.contentType).pipe(take(1)).toPromise();
     if (!contentType) {
-      throw new Error(`Template misconfiguration. No such template for '${contentEntry.contentType}'`);
+      throw new Error(`Template misconfiguration. No such template for '${documentContext.contentType}'`);
     }
 
     const template = await this.contentTemplateService.getTemplate(contentType.template).pipe(take(1)).toPromise();
-    const documentContextData = this.generateDocumentContext(contentType, contentEntry);
-    const module = await this.createModule(documentContextData);
+    const module = await this.createModule(documentContext);
     const factories = await this.compiler.compileModuleAndAllComponentsAsync(module);
     const factory = factories.componentFactories.filter(cf => cf.selector === template.selector)[0];
     if (!factory) {
       throw new Error(`Could not find any template for <${template.selector}>`);
     }
 
-    this.setTitle(contentEntry.title);
     return viewContainer.createComponent(factory);
-  }
-
-  private generateDocumentContext(contentType: ContentType, contentEntry: ContentEntry) {
-    return {
-      id: contentEntry.id,
-      contentType: contentEntry.contentType,
-      data: contentEntry.data,
-      title: contentEntry.title,
-      url: contentType.standalone
-        ? [contentEntry.url.root, contentEntry.url.path].join('/')
-        : null,
-      revision: contentEntry.revision,
-      status: contentEntry.status,
-      tags: contentEntry.tags,
-      created: (contentEntry.createdAt as firebase.firestore.Timestamp).toDate(),
-      updated: (contentEntry.updatedAt as firebase.firestore.Timestamp).toDate(),
-      published: !!contentEntry.publishTime
-        ? (contentEntry.publishTime as firebase.firestore.Timestamp).toDate()
-        : null,
-    } as TanamDocumentContext;
   }
 
   private async createModule(documentContext: TanamDocumentContext) {
@@ -117,7 +95,7 @@ export class DynamicPageService {
 
   private createComponent(template: ContentTemplate, documentContext: TanamDocumentContext) {
     const dynamicClass = class {
-      document = documentContext;
+      context = documentContext;
     };
 
     return Component({
